Address CodeRabbit round 3: extract FormResponseContext, unify submit paths on raw values
- FormResponseContext (+ types, hook) moved to its own module so
  PageRenderer/SinglePageForm no longer import FormRenderer and close a
  circular module chain; FormRenderer re-exports for existing importers
- The <form> submit path now routes through submitCurrentValues, the
  same code path page navigation uses, so both persist an identical
  raw-values payload. Unifying on the resolver-parsed output instead
  would break locked decision §9.5: the visibility-aware schema omits
  hidden fields' keys, so parsed output would wipe hidden values from
  the store on navigation — values must survive until the single
  submit-time strip in handleFormComplete (step 5)
