+++ packages/ui/src/hooks/useFormSubmission.ts
@@ -20,7 +20,15 @@ export const useFormSubmission = (
 ) => {
   const { getValues } = methods;
 
-  // Handle form submission with validation
+  // Handle form submission with validation.
+  //
+  // `data` is deliberately the RAW react-hook-form values, not the
+  // resolver-parsed output: the visibility-aware page schema omits hidden
+  // fields' keys, so persisting parsed output would wipe a hidden field's
+  // typed value from the store on every navigation. Policy is keep-while-
+  // filling, strip at final submit (strategy doc §9.5) — the store mirrors
+  // raw input (same shape useStoreSync writes) and normalization/stripping
+  // happens once, at handleFormComplete.
   const handleSubmit = useCallback(async (data: Record<string, any>) => {
     try {
       // Validate the data using our schema
